REFACTOR: Rename `WireTypes` to `Ports` in abstract wiring diagrams.
Neither nomenclature for the objects of a wiring diagram is perfect,
but I think that `Ports` is clearer. There is some possibility of
confusing `Ports`, a finite list of "port values" or "port types",
with `Port`, a data structure used to describe the source or target
of a `Wire` that implicitly *references* a port value. In particular,
`Ports` is not a list of `Port` objects! Even so, I'll stick with
this nomenclature for now.