BUG: Always parse numbers as values, not references, in JSON S-exprs.
Terms like the following now deserialize properly:

mcopy(A::Ob, n::Int)::Hom(A,opow(A,n))

However, this solution is not ideal because it is no longer possible
to have numeric references. More generally, it is still possible to have
ambiguity between string (or symbolic) values and references. A more robust
solution is to use special objects {"$ref": ...} for references, as in JSON
Schema. On the other hand, that would lead to a more verbose representation.
For now, I'm sticking with the existing encoding.