Stop using deprecated function current_module().
The official advice to use @__MODULE__ instead is not enough. Instead,
we have to grab the current module using the implicit macro parameter
`__module__` at the toplevel, then pass this module down to where it is
needed.

https://github.com/JuliaLang/julia/issues/27835

The messiness of all this is yet more reason to do something about #15.