REFACTOR: Make computing pullbacks/pushouts as composites be explicit.
This change gives us more control. For example, pullbacks of C-sets can
be computed in two ways:

1. Using the pointwise formula for pullbacks in FinSet, computed as
composites of products and equalizers in FinSet.

2. As a composite of products and equalizers in C-Set, computed using
the pointwise formula that reduces them to respectively product and
equalizers in FinSet.

Before this commit, we were inadvertently doing (2), but (1) should be a
bit more efficient, so that's what we're doing now.