REFACTOR: Use `*` instead of `⋆` for horizontal composition.
Besides the fact that the asterisk is a more standard notation for
horizontal composition of natural transformations (used, for example,
Riehl's textbook), this change also gives the horizontal composition in
the double category of V-matrices, which is generalized matrix
multiplication, the same notation as ordinary matrix multiplication in
Julia. This double category is not yet implemented, however.